Revert "copied files"
