--- Singular.xcframework/ios-arm64/Singular.framework/Headers/Attributes.h
@@ -1,41 +1,0 @@
-#ifndef __ATTRIBUTES_H__
-#define __ATTRIBUTES_H__
-
-#define ATTRIBUTE_SNG_ATTR_FROM_DATE @"sng_attr_from_date"
-#define ATTRIBUTE_SNG_ATTR_TO_DATE @"sng_attr_to_date"
-#define ATTRIBUTE_SNG_ATTR_ACHIEVEMENT_ID @"sng_attr_achievement_id"
-#define ATTRIBUTE_SNG_ATTR_CONTENT @"sng_attr_content"
-#define ATTRIBUTE_SNG_ATTR_CONTENT_ID @"sng_attr_content_id"
-#define ATTRIBUTE_SNG_ATTR_CONTENT_LIST @"sng_attr_content_list"
-#define ATTRIBUTE_SNG_ATTR_CONTENT_TYPE @"sng_attr_content_type"
-#define ATTRIBUTE_SNG_ATTR_COUPON_CODE @"sng_attr_coupon_code"
-#define ATTRIBUTE_SNG_ATTR_DEEP_LINK @"sng_attr_deep_link"
-#define ATTRIBUTE_SNG_ATTR_EVENT_END @"sng_attr_event_end"
-#define ATTRIBUTE_SNG_ATTR_EVENT_START @"sng_attr_event_start"
-#define ATTRIBUTE_SNG_ATTR_HOTEL_SCORE @"sng_attr_hotel_score"
-#define ATTRIBUTE_SNG_ATTR_ITEM_DESCRIPTION @"sng_attr_item_description"
-#define ATTRIBUTE_SNG_ATTR_ITEM_PRICE @"sng_attr_item_price"
-#define ATTRIBUTE_SNG_ATTR_LATITUDE @"sng_attr_latitude"
-#define ATTRIBUTE_SNG_ATTR_LEVEL @"sng_attr_level"
-#define ATTRIBUTE_SNG_ATTR_LOCATION @"sng_attr_location"
-#define ATTRIBUTE_SNG_ATTR_LOCATION_ADDRESS_COUNTRY @"sng_attr_location_address_country"
-#define ATTRIBUTE_SNG_ATTR_LOCATION_ADDRESS_REGION_OR_PROVINCE @"sng_attr_location_address_region_or_province"
-#define ATTRIBUTE_SNG_ATTR_LOCATION_ADDRESS_STREET @"sng_attr_location_address_street"
-#define ATTRIBUTE_SNG_ATTR_LONGITUDE @"sng_attr_longitude"
-#define ATTRIBUTE_SNG_ATTR_MAX @"sng_attr_max"
-#define ATTRIBUTE_SNG_ATTR_NEW_VERSION @"sng_attr_new_version"
-#define ATTRIBUTE_SNG_ATTR_ORIGIN @"sng_attr_origin"
-#define ATTRIBUTE_SNG_ATTR_PAYMENT_INFO_AVAILABLE @"sng_attr_payment_info_available"
-#define ATTRIBUTE_SNG_ATTR_QUANTITY @"sng_attr_quantity"
-#define ATTRIBUTE_SNG_ATTR_RATING @"sng_attr_rating"
-#define ATTRIBUTE_SNG_ATTR_REGISTRATION_METHOD @"sng_attr_registration_method"
-#define ATTRIBUTE_SNG_ATTR_REVIEW_TEXT @"sng_attr_review_text"
-#define ATTRIBUTE_SNG_ATTR_SCORE @"sng_attr_score"
-#define ATTRIBUTE_SNG_ATTR_SEARCH_STRING @"sng_attr_search_string"
-#define ATTRIBUTE_SNG_ATTR_SUBSCRIPTION_ID @"sng_attr_subscription_id"
-#define ATTRIBUTE_SNG_ATTR_SUCCESS @"sng_attr_success"
-#define ATTRIBUTE_SNG_ATTR_TRANSACTION_ID @"sng_attr_transaction_id"
-#define ATTRIBUTE_SNG_ATTR_TUTORIAL_ID @"sng_attr_tutorial_id"
-#define ATTRIBUTE_SNG_ATTR_VALID @"sng_attr_valid"
-
-#endif
