--- Singular.xcframework/ios-arm64/Singular.framework/Headers/Events.h
@@ -1,28 +1,0 @@
-#ifndef __EVENTS_H__
-#define __EVENTS_H__
-
-#define EVENT_SNG_RATE @"sng_rate"
-#define EVENT_SNG_SPENT_CREDITS @"sng_spent_credits"
-#define EVENT_SNG_TUTORIAL_COMPLETE @"sng_tutorial_complete"
-#define EVENT_SNG_LOGIN @"sng_login"
-#define EVENT_SNG_START_TRIAL @"sng_start_trial"
-#define EVENT_SNG_SUBSCRIBE @"sng_subscribe"
-#define EVENT_SNG_BOOK @"sng_book"
-#define EVENT_SNG_CONTENT_VIEW_LIST @"sng_content_view_list"
-#define EVENT_SNG_INVITE @"sng_invite"
-#define EVENT_SNG_SHARE @"sng_share"
-#define EVENT_SNG_SUBMIT_APPLICATION @"sng_submit_application"
-#define EVENT_SNG_UPDATE @"sng_update"
-#define EVENT_SNG_ECOMMERCE_PURCHASE @"sng_ecommerce_purchase"
-#define EVENT_SNG_VIEW_CART @"sng_view_cart"
-#define EVENT_SNG_ACHIEVEMENT_UNLOCKED @"sng_achievement_unlocked"
-#define EVENT_SNG_ADD_PAYMENT_INFO @"sng_add_payment_info"
-#define EVENT_SNG_ADD_TO_CART @"sng_add_to_cart"
-#define EVENT_SNG_ADD_TO_WISHLIST @"sng_add_to_wishlist"
-#define EVENT_SNG_CHECKOUT_INITIATED @"sng_checkout_initiated"
-#define EVENT_SNG_COMPLETE_REGISTRATION @"sng_complete_registration"
-#define EVENT_SNG_CONTENT_VIEW @"sng_content_view"
-#define EVENT_SNG_LEVEL_ACHIEVED @"sng_level_achieved"
-#define EVENT_SNG_SEARCH @"sng_search"
-
-#endif
