+++ Singular.xcframework/ios-arm64_armv7_armv7s/Singular.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module Singular {
+  umbrella header "Singular.h"
+
+  export *
+  module * { export * }
+}
