+++ Singular.xcframework/ios-arm64_i386_x86_64-simulator/Singular.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module Singular {
+  umbrella header "Singular.h"
+
+  export *
+  module * { export * }
+}
