--- Singular.xcframework/ios-arm64_x86_64-simulator/Singular.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module Singular {
-  umbrella header "Singular.h"
-
-  export *
-  module * { export * }
-}
